wip: remover carrinho antes de destruir produto

+++ app/controllers/i_produtos_controller.rb
@@ -34,13 +34,20 @@ class IProdutosController < ApplicationController
     end
   end
 
-  def destroy
+def destroy
+  if @i_produto.status != "inativo"
+    redirect_to i_produtos_url, alert: "Só é possível excluir produtos com status 'inativo'."
+  else
+    @i_produto.i_itens_carrinhos.destroy_all
+
     if @i_produto.destroy
       redirect_to i_produtos_url, notice: t('messages.deleted_successfully')
     else
       redirect_to i_produtos_url, alert: t('messages.delete_failed_due_to_dependencies')
-    end   
+    end
   end
+end
+
 
 
   private
